feat(data exploration): convert funnel correlation events table (#14977)
