+++ frontend/src/scenes/funnels/funnelLogic.test.ts
@@ -1,25 +1,13 @@
-import { DEFAULT_EXCLUDED_PERSON_PROPERTIES, funnelLogic } from './funnelLogic'
-import { MOCK_DEFAULT_TEAM, MOCK_TEAM_ID } from 'lib/api.mock'
-import { expectLogic, partial } from 'kea-test-utils'
+import { funnelLogic } from './funnelLogic'
+import { MOCK_TEAM_ID } from 'lib/api.mock'
+import { expectLogic } from 'kea-test-utils'
 import { initKeaTests } from '~/test/init'
 import { preflightLogic } from 'scenes/PreflightCheck/preflightLogic'
 import { eventUsageLogic } from 'lib/utils/eventUsageLogic'
 import { insightLogic } from 'scenes/insights/insightLogic'
-import {
-    AvailableFeature,
-    CorrelationConfigType,
-    FunnelCorrelation,
-    FunnelCorrelationResultsType,
-    FunnelCorrelationType,
-    FunnelsFilterType,
-    FunnelVizType,
-    InsightLogicProps,
-    InsightShortId,
-    InsightType,
-} from '~/types'
+import { AvailableFeature, InsightLogicProps, InsightShortId, InsightType } from '~/types'
 import { teamLogic } from 'scenes/teamLogic'
 import { userLogic } from 'scenes/userLogic'
-import { groupPropertiesModel } from '~/models/groupPropertiesModel'
 import { router } from 'kea-router'
 import { urls } from 'scenes/urls'
 import { useMocks } from '~/mocks/jest'
@@ -29,86 +17,7 @@ import { openPersonsModal } from 'scenes/trends/persons-modal/PersonsModal'
 
 jest.mock('scenes/trends/persons-modal/PersonsModal')
 
-const Insight12 = '12' as InsightShortId
 const Insight123 = '123' as InsightShortId
-
-export const mockInsight = {
-    id: Insight123,
-    short_id: 'SvoU2bMC',
-    name: null,
-    filters: {
-        breakdown: null,
-        breakdown_type: null,
-        display: 'FunnelViz',
-        events: [
-            {
-                id: '$pageview',
-                type: 'events',
-                order: 0,
-                name: '$pageview',
-                custom_name: null,
-                math: null,
-                math_property: null,
-                properties: [],
-            },
-            {
-                id: '$pageview',
-                type: 'events',
-                order: 1,
-                name: '$pageview',
-                custom_name: null,
-                math: null,
-                math_property: null,
-                properties: [],
-            },
-            {
-                id: '$pageview',
-                type: 'events',
-                order: 2,
-                name: '$pageview',
-                custom_name: null,
-                math: null,
-                math_property: null,
-                properties: [],
-            },
-            {
-                id: '$pageview',
-                type: 'events',
-                order: 3,
-                name: '$pageview',
-                custom_name: null,
-                math: null,
-                math_property: null,
-                properties: [],
-            },
-        ],
-        funnel_from_step: 0,
-        funnel_to_step: 1,
-        funnel_viz_type: 'steps',
-        insight: 'FUNNELS',
-        layout: 'vertical',
-    },
-    order: null,
-    deleted: false,
-    dashboard: null,
-    layouts: {},
-    color: null,
-    last_refresh: null,
-    result: null,
-    created_at: '2021-09-22T18:22:20.036153Z',
-    description: null,
-    updated_at: '2021-09-22T19:03:49.322258Z',
-    tags: [],
-    favorited: false,
-    saved: false,
-    created_by: {
-        id: 1,
-        uuid: '017c0441-bcb2-0000-bccf-dfc24328c5f3',
-        distinct_id: 'fM7b6ZFi8MOssbkDI55ot8tMY2hkzrHdRy1qERa6rCK',
-        first_name: 'Alex',
-        email: 'alex@posthog.com',
-    },
-}
 
 const funnelResults = [
     {
@@ -136,200 +45,24 @@ const funnelResults = [
 
 describe('funnelLogic', () => {
     let logic: ReturnType<typeof funnelLogic.build>
-    let correlationConfig: CorrelationConfigType = {}
 
     beforeEach(() => {
         useAvailableFeatures([AvailableFeature.CORRELATION_ANALYSIS, AvailableFeature.GROUP_ANALYTICS])
         useMocks({
             get: {
-                '/api/projects/@current': () => [
-                    200,
-                    {
-                        ...MOCK_DEFAULT_TEAM,
-                        correlation_config: correlationConfig,
-                    },
-                ],
-                '/api/projects/:team/insights/': (req) => {
-                    if (req.url.searchParams.get('saved')) {
-                        return [
-                            200,
-                            {
-                                results: funnelResults,
-                            },
-                        ]
-                    }
-                    const shortId = req.url.searchParams.get('short_id') || ''
-                    if (shortId === '500') {
-                        return [500, { status: 0, detail: 'error from the API' }]
-                    }
-                    return [
-                        200,
-                        {
-                            results: [mockInsight],
-                        },
-                    ]
+                '/api/projects/:team/insights/': {
+                    results: [{}],
                 },
-                '/api/projects/:team/insights/trend/': { results: ['trends result from api'] },
+                '/api/projects/:team/insights/:id/': {},
                 '/api/projects/:team/groups_types/': [],
-                '/some/people/url': { results: [{ people: [] }] },
-                '/api/projects/:team/persons/funnel': { results: [], next: null },
-                '/api/projects/:team/persons/properties': [
-                    { name: 'some property', count: 20 },
-                    { name: 'another property', count: 10 },
-                    { name: 'third property', count: 5 },
-                ],
-                '/api/projects/:team/groups/property_definitions': {
-                    '0': [
-                        { name: 'industry', count: 2 },
-                        { name: 'name', count: 1 },
-                    ],
-                    '1': [{ name: 'name', count: 1 }],
-                },
-            },
-            patch: {
-                '/api/projects/:id': (req) => [
-                    200,
-                    {
-                        ...MOCK_DEFAULT_TEAM,
-                        correlation_config: {
-                            ...correlationConfig,
-                            excluded_person_property_names: (req.body as any)?.correlation_config
-                                ?.excluded_person_property_names,
-                        },
-                    },
-                ],
             },
             post: {
-                '/api/projects/:team/insights/': (req) => [
-                    200,
-                    { id: 12, short_id: Insight12, ...((req.body as any) || {}) },
-                ],
-                '/api/projects/:team/insights/:id/viewed': [201],
                 '/api/projects/:team/insights/funnel/': {
-                    is_cached: true,
-                    last_refresh: '2021-09-16T13:41:41.297295Z',
                     result: funnelResults,
-                    type: 'Funnel',
-                },
-                '/api/projects/:team/insights/funnel/correlation': (req) => {
-                    const data = req.body as any
-                    if (data?.funnel_correlation_type === 'properties') {
-                        const excludePropertyFromProjectNames = data?.funnel_correlation_exclude_names || []
-                        const includePropertyNames = data?.funnel_correlation_names || []
-                        return [
-                            200,
-                            {
-                                is_cached: true,
-                                last_refresh: '2021-09-16T13:41:41.297295Z',
-                                result: {
-                                    events: [
-                                        {
-                                            event: { event: 'some property' },
-                                            success_count: 1,
-                                            failure_count: 1,
-                                            odds_ratio: 1,
-                                            correlation_type: 'success',
-                                        },
-                                        {
-                                            event: { event: 'another property' },
-                                            success_count: 1,
-                                            failure_count: 1,
-                                            odds_ratio: 1,
-                                            correlation_type: 'failure',
-                                        },
-                                    ]
-                                        .filter(
-                                            (correlation) =>
-                                                includePropertyNames.includes('$all') ||
-                                                includePropertyNames.includes(correlation.event.event)
-                                        )
-                                        .filter(
-                                            (correlation) =>
-                                                !excludePropertyFromProjectNames.includes(correlation.event.event)
-                                        ),
-                                },
-                                type: 'Funnel',
-                            },
-                        ]
-                    } else if (data?.funnel_correlation_type === 'events') {
-                        return [
-                            200,
-                            {
-                                is_cached: true,
-                                last_refresh: '2021-09-16T13:41:41.297295Z',
-                                result: {
-                                    events: [
-                                        {
-                                            event: { event: 'some event' },
-                                            success_count: 1,
-                                            failure_count: 1,
-                                            odds_ratio: 1,
-                                            correlation_type: 'success',
-                                        },
-                                        {
-                                            event: { event: 'another event' },
-                                            success_count: 1,
-                                            failure_count: 1,
-                                            odds_ratio: 1,
-                                            correlation_type: 'failure',
-                                        },
-                                    ],
-                                },
-                                type: 'Funnel',
-                            },
-                        ]
-                    } else if (data?.funnel_correlation_type === 'event_with_properties') {
-                        const targetEvent = data?.funnel_correlation_event_names[0]
-                        const excludedProperties = data?.funnel_correlation_event_exclude_property_names
-                        return [
-                            200,
-                            {
-                                result: {
-                                    events: [
-                                        {
-                                            success_count: 1,
-                                            failure_count: 0,
-                                            odds_ratio: 29,
-                                            correlation_type: 'success',
-                                            event: { event: `some event::name::Hester` },
-                                        },
-                                        {
-                                            success_count: 1,
-                                            failure_count: 0,
-                                            odds_ratio: 29,
-                                            correlation_type: 'success',
-                                            event: { event: `some event::Another name::Alice` },
-                                        },
-                                        {
-                                            success_count: 1,
-                                            failure_count: 0,
-                                            odds_ratio: 25,
-                                            correlation_type: 'success',
-                                            event: { event: `another event::name::Aloha` },
-                                        },
-                                        {
-                                            success_count: 1,
-                                            failure_count: 0,
-                                            odds_ratio: 25,
-                                            correlation_type: 'success',
-                                            event: { event: `another event::Another name::Bob` },
-                                        },
-                                    ].filter(
-                                        (record) =>
-                                            record.event.event.split('::')[0] === targetEvent &&
-                                            !excludedProperties.includes(record.event.event.split('::')[1])
-                                    ),
-                                    last_refresh: '2021-11-05T09:26:16.175923Z',
-                                    is_cached: false,
-                                },
-                            },
-                        ]
-                    }
                 },
             },
         })
         initKeaTests(false)
-        window.POSTHOG_APP_CONTEXT = undefined // to force API request to /api/project/@current
     })
 
     const defaultProps: InsightLogicProps = {
@@ -622,441 +355,11 @@ describe('funnelLogic', () => {
         })
     })
 
-    describe('selectors', () => {
-        beforeEach(async () => {
-            await initFunnelLogic()
-        })
-        describe('Correlation Names parsing', () => {
-            const basicFunnelRecord: FunnelCorrelation = {
-                event: { event: '$pageview::bzzz', properties: {}, elements: [] },
-                odds_ratio: 1,
-                correlation_type: FunnelCorrelationType.Success,
-                success_count: 1,
-                failure_count: 1,
-                success_people_url: '/some/people/url',
-                failure_people_url: '/some/people/url',
-                result_type: FunnelCorrelationResultsType.Events,
-            }
-            it('chooses the correct name based on Event type', async () => {
-                const result = logic.values.parseDisplayNameForCorrelation(basicFunnelRecord)
-                expect(result).toEqual({
-                    first_value: '$pageview::bzzz',
-                    second_value: undefined,
-                })
-            })
-
-            it('chooses the correct name based on Property type', async () => {
-                const result = logic.values.parseDisplayNameForCorrelation({
-                    ...basicFunnelRecord,
-                    result_type: FunnelCorrelationResultsType.Properties,
-                })
-                expect(result).toEqual({
-                    first_value: '$pageview',
-                    second_value: 'bzzz',
-                })
-            })
-
-            it('chooses the correct name based on EventWithProperty type', async () => {
-                const result = logic.values.parseDisplayNameForCorrelation({
-                    ...basicFunnelRecord,
-                    result_type: FunnelCorrelationResultsType.EventWithProperties,
-                    event: {
-                        event: '$pageview::library::1.2',
-                        properties: { random: 'x' },
-                        elements: [],
-                    },
-                })
-                expect(result).toEqual({
-                    first_value: 'library',
-                    second_value: '1.2',
-                })
-            })
-
-            it('handles autocapture events on EventWithProperty type', async () => {
-                const result = logic.values.parseDisplayNameForCorrelation({
-                    ...basicFunnelRecord,
-                    result_type: FunnelCorrelationResultsType.EventWithProperties,
-                    event: {
-                        event: '$autocapture::elements_chain::xyz_elements_a.link*',
-                        properties: { $event_type: 'click' },
-                        elements: [
-                            {
-                                tag_name: 'a',
-                                href: '#',
-                                attributes: { blah: 'https://example.com' },
-                                nth_child: 0,
-                                nth_of_type: 0,
-                                order: 0,
-                                text: 'bazinga',
-                            },
-                        ],
-                    },
-                })
-                expect(result).toEqual({
-                    first_value: 'clicked link with text "bazinga"',
-                    second_value: undefined,
-                })
-            })
-
-            it('handles autocapture events without elements_chain on EventWithProperty type', async () => {
-                const result = logic.values.parseDisplayNameForCorrelation({
-                    ...basicFunnelRecord,
-                    result_type: FunnelCorrelationResultsType.EventWithProperties,
-                    event: {
-                        event: '$autocapture::library::1.2',
-                        properties: { random: 'x' },
-                        elements: [],
-                    },
-                })
-                expect(result).toEqual({
-                    first_value: 'library',
-                    second_value: '1.2',
-                })
-            })
-        })
-    })
-
-    describe('funnel correlation matrix', () => {
-        beforeEach(async () => {
-            await initFunnelLogic()
-        })
-        it('Selecting a record returns appropriate values', async () => {
-            await expectLogic(logic, () =>
-                logic.actions.setFunnelCorrelationDetails({
-                    event: { event: 'some event', elements: [], properties: {} },
-                    success_people_url: '',
-                    failure_people_url: '',
-                    success_count: 2,
-                    failure_count: 4,
-                    odds_ratio: 3,
-                    correlation_type: FunnelCorrelationType.Success,
-                    result_type: FunnelCorrelationResultsType.Events,
-                })
-            ).toMatchValues({
-                correlationMatrixAndScore: {
-                    correlationScore: expect.anything(),
-                    correlationScoreStrength: 'weak',
-                    truePositive: 2,
-                    falsePositive: 2,
-                    trueNegative: 11,
-                    falseNegative: 4,
-                },
-            })
-
-            expect(logic.values.correlationMatrixAndScore.correlationScore).toBeCloseTo(0.204)
-        })
-    })
-
-    describe('funnel correlation properties', () => {
-        const props = { dashboardItemId: Insight123, syncWithUrl: true }
-
-        it('Selecting all properties returns expected result', async () => {
-            await initFunnelLogic(props)
-            await expectLogic(logic, () => logic.actions.setPropertyNames(logic.values.allProperties))
-                .toFinishListeners()
-                .toMatchValues({
-                    propertyCorrelations: {
-                        events: [
-                            {
-                                event: { event: 'some property' },
-                                success_count: 1,
-                                failure_count: 1,
-                                odds_ratio: 1,
-                                correlation_type: 'success',
-                                result_type: FunnelCorrelationResultsType.Properties,
-                            },
-                            {
-                                event: { event: 'another property' },
-                                success_count: 1,
-                                failure_count: 1,
-                                odds_ratio: 1,
-                                correlation_type: 'failure',
-                                result_type: FunnelCorrelationResultsType.Properties,
-                            },
-                        ],
-                    },
-                })
-        })
-
-        it('Deselecting all returns empty result', async () => {
-            await initFunnelLogic(props)
-            await expectLogic(logic, () => logic.actions.setPropertyNames([]))
-                .toDispatchActions(logic, ['loadPropertyCorrelationsSuccess'])
-                .toMatchValues({
-                    propertyCorrelations: {
-                        events: [],
-                    },
-                })
-        })
-
-        // TODO: loading of property correlations is now dependent on the table being shown in react
-        it.skip('are updated when results are loaded, when steps visualisation set', async () => {
-            await initFunnelLogic(props)
-            const filters = {
-                insight: InsightType.FUNNELS,
-                funnel_viz_type: FunnelVizType.Steps,
-            }
-            await router.actions.push(urls.insightNew(filters))
-
-            await expectLogic(logic)
-                .toFinishAllListeners()
-                .toMatchValues({
-                    steps: [
-                        { action_id: '$pageview', count: 19, name: '$pageview', order: 0, type: 'events' },
-                        { action_id: '$pageview', count: 7, name: '$pageview', order: 1, type: 'events' },
-                        { action_id: '$pageview', count: 4, name: '$pageview', order: 2, type: 'events' },
-                    ],
-                    propertyCorrelations: {
-                        events: [
-                            {
-                                event: { event: 'some property' },
-                                success_count: 1,
-                                failure_count: 1,
-                                odds_ratio: 1,
-                                correlation_type: 'success',
-                                result_type: FunnelCorrelationResultsType.Properties,
-                            },
-                            {
-                                event: { event: 'another property' },
-                                success_count: 1,
-                                failure_count: 1,
-                                odds_ratio: 1,
-                                correlation_type: 'failure',
-                                result_type: FunnelCorrelationResultsType.Properties,
-                            },
-                        ],
-                    },
-                })
-        })
-        it('are not updated when results are loaded, when steps visualisation set, with one funnel step', async () => {
-            await initFunnelLogic(props)
-
-            await expectLogic(logic, () => {
-                logic.actions.loadResultsSuccess({
-                    filters: {
-                        insight: InsightType.FUNNELS,
-                        funnel_viz_type: FunnelVizType.Steps,
-                    } as FunnelsFilterType,
-                    result: [{ action_id: 'some event', order: 0 }],
-                })
-            })
-                .toFinishListeners()
-                .toMatchValues({
-                    steps: [{ action_id: 'some event', order: 0 }],
-                    propertyCorrelations: {
-                        events: [],
-                    },
-                    correlations: {
-                        events: [],
-                    },
-                })
-        })
-        it('are not triggered when results are loaded, when trends visualisation set', async () => {
-            await initFunnelLogic(props)
-            await expectLogic(logic, () => {
-                logic.actions.loadResultsSuccess({
-                    filters: {
-                        insight: InsightType.FUNNELS,
-                        funnel_viz_type: FunnelVizType.Trends,
-                    } as FunnelsFilterType,
-                })
-            }).toNotHaveDispatchedActions(['loadEventCorrelations', 'loadPropertyCorrelations'])
-        })
-
-        it('triggers update to correlation list when property excluded from project', async () => {
-            userLogic.mount()
-            await initFunnelLogic(props)
-            //
-            // // Make sure we have loaded the team already
-            // await expectLogic(teamLogic, () => teamLogic.actions.loadCurrentTeam()).toFinishAllListeners()
-
-            await expectLogic(logic, () => {
-                logic.actions.setPropertyNames(logic.values.allProperties)
-                logic.actions.loadResultsSuccess({ filters: { insight: InsightType.FUNNELS } })
-                logic.actions.excludePropertyFromProject('another property')
-            })
-                .toFinishAllListeners()
-                .toMatchValues({
-                    propertyNames: ['some property', 'third property'],
-                    excludedPropertyNames: DEFAULT_EXCLUDED_PERSON_PROPERTIES.concat(['another property']),
-                    allProperties: ['some property', 'third property'],
-                })
-
-            expect(logic.values.propertyCorrelationValues).toEqual([
-                {
-                    event: { event: 'some property' },
-                    success_count: 1,
-                    failure_count: 1,
-                    odds_ratio: 1,
-                    correlation_type: 'success',
-                    result_type: FunnelCorrelationResultsType.Properties,
-                },
-            ])
-        })
-
-        it('isPropertyExcludedFromProject returns true initially, then false when excluded, and is persisted to team config', async () => {
-            await initFunnelLogic(props)
-
-            expect(logic.values.isPropertyExcludedFromProject('some property')).toBe(false)
-
-            await expectLogic(logic, () =>
-                logic.actions.excludePropertyFromProject('some property')
-            ).toFinishAllListeners()
-
-            expect(logic.values.isPropertyExcludedFromProject('some property')).toBe(true)
-
-            await expectLogic(teamLogic).toMatchValues({
-                currentTeam: partial({
-                    correlation_config: {
-                        excluded_person_property_names: DEFAULT_EXCLUDED_PERSON_PROPERTIES.concat(['some property']),
-                    },
-                }),
-            })
-
-            // Also make sure that excluding the property again doesn't double
-            // up on the config list
-            await expectLogic(logic, () =>
-                logic.actions.excludePropertyFromProject('some property')
-            ).toFinishAllListeners()
-
-            await expectLogic(teamLogic).toMatchValues({
-                currentTeam: partial({
-                    correlation_config: {
-                        excluded_person_property_names: DEFAULT_EXCLUDED_PERSON_PROPERTIES.concat(['some property']),
-                    },
-                }),
-            })
-        })
-
-        it('loads property exclude list from Project settings', async () => {
-            correlationConfig = { excluded_person_property_names: ['some property'] }
-            await initFunnelLogic(props)
-
-            await expectLogic(teamLogic).toMatchValues({
-                currentTeam: partial({
-                    correlation_config: { excluded_person_property_names: ['some property'] },
-                }),
-            })
-
-            await expectLogic(logic, () => {
-                logic.actions.setPropertyNames(logic.values.allProperties)
-                logic.actions.loadResultsSuccess({ filters: { insight: InsightType.FUNNELS } })
-            })
-                .toFinishAllListeners()
-                .toMatchValues({
-                    propertyCorrelations: {
-                        events: [
-                            {
-                                event: { event: 'another property' },
-                                success_count: 1,
-                                failure_count: 1,
-                                odds_ratio: 1,
-                                correlation_type: 'failure',
-                                result_type: FunnelCorrelationResultsType.Properties,
-                            },
-                        ],
-                    },
-                })
-        })
-
-        // TODO: loading of correlations is now dependent on the table being shown in react
-        it.skip('loads event exclude list from Project settings', async () => {
-            correlationConfig = { excluded_event_names: ['some event'] }
-            await initFunnelLogic(props)
-
-            await expectLogic(teamLogic).toMatchValues({
-                currentTeam: partial({
-                    correlation_config: { excluded_event_names: ['some event'] },
-                }),
-            })
-
-            const filters = {
-                insight: InsightType.FUNNELS,
-                funnel_viz_type: FunnelVizType.Steps,
-            }
-            await router.actions.push(urls.insightNew(filters))
-
-            await expectLogic(logic)
-                .toFinishAllListeners()
-                .toMatchValues({
-                    correlationValues: [
-                        {
-                            event: { event: 'another event' },
-                            success_count: 1,
-                            failure_count: 1,
-                            odds_ratio: 1,
-                            correlation_type: 'failure',
-                            result_type: FunnelCorrelationResultsType.Events,
-                        },
-                    ],
-                })
-        })
-
-        it('loads event property exclude list from Project settings', async () => {
-            correlationConfig = { excluded_event_property_names: ['name'] }
-            await initFunnelLogic(props)
-
-            await expectLogic(teamLogic).toMatchValues({
-                currentTeam: partial({
-                    correlation_config: { excluded_event_property_names: ['name'] },
-                }),
-            })
-
-            await expectLogic(logic, () => {
-                logic.actions.loadEventWithPropertyCorrelations('some event')
-            })
-                .toDispatchActions(logic, ['loadEventWithPropertyCorrelationsSuccess'])
-                .toFinishListeners()
-                .toMatchValues({
-                    eventWithPropertyCorrelations: {
-                        'some event': [
-                            {
-                                event: { event: 'some event::Another name::Alice' },
-                                success_count: 1,
-                                failure_count: 0,
-                                odds_ratio: 29,
-                                correlation_type: 'success',
-                                result_type: FunnelCorrelationResultsType.EventWithProperties,
-                            },
-                        ],
-                    },
-                })
-        })
-
-        // TODO: fix this test
-        it.skip('Selecting all group properties selects correct properties', async () => {
-            await initFunnelLogic(props)
-
-            groupPropertiesModel.mount()
-            groupPropertiesModel.actions.loadAllGroupProperties()
-            await expectLogic(groupPropertiesModel).toDispatchActions(['loadAllGroupPropertiesSuccess'])
-
-            const filters = {
-                insight: InsightType.FUNNELS,
-                funnel_viz_type: FunnelVizType.Steps,
-            }
-            await router.actions.push(urls.insightNew(filters))
-
-            await expectLogic(logic, () => logic.actions.setFilters({ aggregation_group_type_index: 0 }))
-                .toFinishAllListeners()
-                .toMatchValues({
-                    allProperties: ['industry', 'name'],
-                    propertyNames: ['industry', 'name'],
-                })
-
-            await expectLogic(logic, () => logic.actions.setFilters({ aggregation_group_type_index: 1 }))
-                .toFinishAllListeners()
-                .toMatchValues({
-                    allProperties: ['name'],
-                    propertyNames: ['name'],
-                })
-        })
-    })
-
     describe('funnel simple vs. advanced mode', () => {
         beforeEach(async () => {
             await initFunnelLogic()
         })
+
         it("toggleAdvancedMode() doesn't trigger a load result", async () => {
             await expectLogic(logic, () => {
                 logic.actions.toggleAdvancedMode()
