+++ frontend/src/scenes/funnels/funnelUtils.ts
@@ -1,4 +1,4 @@
-import { clamp } from 'lib/utils'
+import { autoCaptureEventToDescription, clamp } from 'lib/utils'
 import {
     FunnelStepRangeEntityFilter,
     FunnelStep,
@@ -16,12 +16,15 @@ import {
     PropertyOperator,
     ElementPropertyFilter,
     PropertyFilterType,
+    FunnelCorrelationResultsType,
+    CorrelationConfigType,
 } from '~/types'
 import { dayjs } from 'lib/dayjs'
 import { combineUrl } from 'kea-router'
 import { FunnelsQuery } from '~/queries/schema'
 import { FunnelLayout } from 'lib/constants'
 import { elementsToAction } from 'scenes/events/createActionFromEvent'
+import { teamLogic } from 'scenes/teamLogic'
 
 /** Chosen via heuristics by eyeballing some values
  * Assuming a normal distribution, then 90% of values are within 1.5 standard deviations of the mean
@@ -521,6 +524,23 @@ export const transformLegacyHiddenLegendKeys = (
     return hiddenLegendKeys
 }
 
+export const parseBreakdownValue = (
+    item: string
+): {
+    breakdown: string
+    breakdown_value: string
+} => {
+    const components = item.split('::')
+    if (components.length === 1) {
+        return { breakdown: components[0], breakdown_value: '' }
+    } else {
+        return {
+            breakdown: components[0],
+            breakdown_value: components[1],
+        }
+    }
+}
+
 export const parseEventAndProperty = (
     event: FunnelCorrelation['event']
 ): {
@@ -560,4 +580,65 @@ export const parseEventAndProperty = (
             ],
         }
     }
+}
+
+export const parseDisplayNameForCorrelation = (
+    record: FunnelCorrelation
+): { first_value: string; second_value?: string } => {
+    let first_value = undefined
+    let second_value = undefined
+    const values = record.event.event.split('::')
+
+    if (record.result_type === FunnelCorrelationResultsType.Events) {
+        first_value = record.event.event
+        return { first_value, second_value }
+    } else if (record.result_type === FunnelCorrelationResultsType.Properties) {
+        first_value = values[0]
+        second_value = values[1]
+        return { first_value, second_value }
+    } else if (values[0] === '$autocapture' && values[1] === 'elements_chain') {
+        // special case for autocapture elements_chain
+        first_value = autoCaptureEventToDescription({
+            ...record.event,
+            event: '$autocapture',
+        }) as string
+        return { first_value, second_value }
+    } else {
+        // FunnelCorrelationResultsType.EventWithProperties
+        // Events here come in the form of event::property::value
+        return { first_value: values[1], second_value: values[2] }
+    }
+}
+
+export const appendToCorrelationConfig = (
+    configKey: keyof CorrelationConfigType,
+    currentValue: string[],
+    configValue: string
+): void => {
+    // Helper to handle updating correlationConfig within the Team model. Only
+    // handles further appending to current values.
+
+    // When we exclude a property, we want to update the config stored
+    // on the current Team/Project.
+    const oldCurrentTeam = teamLogic.values.currentTeam
+
+    // If we haven't actually retrieved the current team, we can't
+    // update the config.
+    if (oldCurrentTeam === null || !currentValue) {
+        console.warn('Attempt to update correlation config without first retrieving existing config')
+        return
+    }
+
+    const oldCorrelationConfig = oldCurrentTeam.correlation_config
+
+    const configList = [...Array.from(new Set(currentValue.concat([configValue])))]
+
+    const correlationConfig = {
+        ...oldCorrelationConfig,
+        [configKey]: configList,
+    }
+
+    teamLogic.actions.updateCurrentTeam({
+        correlation_config: correlationConfig,
+    })
 }
