+++ frontend/src/scenes/insights/views/Funnels/FunnelCorrelation.tsx
@@ -1,26 +1,32 @@
 import { useMountedLogic, useValues } from 'kea'
-import { funnelLogic } from 'scenes/funnels/funnelLogic'
-import './FunnelCorrelation.scss'
-import { AvailableFeature } from '~/types'
+
 import { insightLogic } from 'scenes/insights/insightLogic'
-import { FunnelCorrelationTable } from './FunnelCorrelationTable'
-import { FunnelPropertyCorrelationTable } from './FunnelPropertyCorrelationTable'
-import { PayGateMini } from 'lib/components/PayGateMini/PayGateMini'
+import { funnelLogic } from 'scenes/funnels/funnelLogic'
 import { funnelDataLogic } from 'scenes/funnels/funnelDataLogic'
+import { funnelCorrelationUsageLogic } from 'scenes/funnels/funnelCorrelationUsageLogic'
+
+import { PayGateMini } from 'lib/components/PayGateMini/PayGateMini'
 import {
     FunnelCorrelationSkewWarning,
     FunnelCorrelationSkewWarningDataExploration,
 } from './FunnelCorrelationSkewWarning'
+import { FunnelCorrelationTable, FunnelCorrelationTableDataExploration } from './FunnelCorrelationTable'
 import { FunnelCorrelationFeedbackForm } from './FunnelCorrelationFeedbackForm'
-import { funnelCorrelationUsageLogic } from 'scenes/funnels/funnelCorrelationUsageLogic'
+import {
+    FunnelPropertyCorrelationTable,
+    FunnelPropertyCorrelationTableDataExploration,
+} from './FunnelPropertyCorrelationTable'
+import { AvailableFeature } from '~/types'
+
+import './FunnelCorrelation.scss'
 
 export const FunnelCorrelation = (): JSX.Element | null => {
-    const { insightProps, isUsingDataExploration } = useValues(insightLogic)
+    const { insightProps, isUsingDataExploration: dx } = useValues(insightLogic)
     const { steps: legacySteps } = useValues(funnelLogic(insightProps))
     const { steps } = useValues(funnelDataLogic(insightProps))
     useMountedLogic(funnelCorrelationUsageLogic(insightProps))
 
-    if (isUsingDataExploration ? steps.length <= 1 : legacySteps.length <= 1) {
+    if (dx ? steps.length <= 1 : legacySteps.length <= 1) {
         return null
     }
 
@@ -29,14 +35,10 @@ export const FunnelCorrelation = (): JSX.Element | null => {
             <h2 className="my-4">Correlation analysis</h2>
             <PayGateMini feature={AvailableFeature.CORRELATION_ANALYSIS}>
                 <div className="funnel-correlation">
-                    {isUsingDataExploration ? (
-                        <FunnelCorrelationSkewWarningDataExploration />
-                    ) : (
-                        <FunnelCorrelationSkewWarning />
-                    )}
-                    {!isUsingDataExploration && <FunnelCorrelationTable />}
+                    {dx ? <FunnelCorrelationSkewWarningDataExploration /> : <FunnelCorrelationSkewWarning />}
+                    {dx ? <FunnelCorrelationTableDataExploration /> : <FunnelCorrelationTable />}
                     <FunnelCorrelationFeedbackForm />
-                    {!isUsingDataExploration && <FunnelPropertyCorrelationTable />}
+                    {dx ? <FunnelPropertyCorrelationTableDataExploration /> : <FunnelPropertyCorrelationTable />}
                 </div>
             </PayGateMini>
         </>
